Refactor: 코드 리팩토링
CustomAnswerRepository,
CustomAnswerRepositoryImpl:
- 코드 정렬
- 안쓰는 기능인 findCategoryId 삭제

CustomMemberAnswerRepository:
- 기능 설명주석 추가

CustomMemberAnswerRepositoryImpl:
- 코드 정렬
- MemberAnswer 테이블에서 categoryId를 삭제하고 join으로 대체

KakaoOAuthDto:
- 생성자 롤백 (카카오 토큰 생성하는 과정에서 생성자가 없어서 오류 발생)

MemberAnswer:
- category_id 삭제 (join으로 대체)

GlobalControllerExceptionHandler:
- 코드 정렬
- 에러 발생시 log.error 로그 메시지 띄우도록 변경

MemberAnswerService:
- 코드 정렬
- saveAnswerList
  1. 카테고리 ID가 db에 없을때 예외 추가
- matchedMemberList
  1. Transactional(readOnly) True 설정(읽기전용)

MemberService
- 코드 정렬
- getKakaoUserInfo
  1. 가져온 카카오 userInfo에 nickname, email, gender, profile_image_url이 없을때도 정상적으로 가입 되도록 해당 값이 없으면 빈 값 ("") 넣도록 변경

+++ src/main/java/com/gacheeGame/service/MemberAnswerService.java
@@ -4,43 +4,38 @@ import com.gacheeGame.dto.AnswerDto;
 import com.gacheeGame.dto.AnswerDto.AnswerList;
 import com.gacheeGame.dto.MemberAnswerDto;
 import com.gacheeGame.dto.MemberAnswerDto.MatchedDto;
-import com.gacheeGame.dto.ResponseDto;
 import com.gacheeGame.entity.MemberAnswer;
 import com.gacheeGame.handler.CustomBadRequestException;
 import com.gacheeGame.mapper.MemberAnswerMapper;
 import com.gacheeGame.repository.AnswerRepository;
+import com.gacheeGame.repository.CategoryRepository;
 import com.gacheeGame.repository.MemberAnswerRepository;
 import com.gacheeGame.repository.MemberRepository;
-import com.gacheeGame.util.JsonUtil;
 import com.gacheeGame.util.SessionUtil;
-import java.util.Arrays;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class MemberAnswerService
-{
+public class MemberAnswerService {
     private final MemberAnswerRepository memberAnswerRepository;
     private final AnswerRepository answerRepository;
     private final MemberRepository memberRepository;
     private final MemberAnswerMapper memberAnswerMapper;
+    private final CategoryRepository categoryRepository;
 
     // 상대방 or 내 답변 가져오기
     @Transactional(readOnly = true)
-    public HashMap<String,Object> memberAnswerList(Long matchedMemberId, Long categoryId)
-    {
-        try{
+    public HashMap<String, Object> memberAnswerList(Long matchedMemberId, Long categoryId) {
+        try {
             HashMap<String, Object> resultMap = new HashMap<>();
             HashMap<String, Object> memberAnswerListMap = new HashMap<>();
 
@@ -50,12 +45,11 @@ public class MemberAnswerService
             //내 답변 목록 조회
             List<AnswerDto.AnswerList> myAnswerList = answerRepository.findAnswerList(memberId, matchedMemberId, categoryId);
 
-            //결과 Map에 저장
+            //내 답변 목록 Map에 저장
             memberAnswerListMap.put("myAnswerList", myAnswerList);
 
             //공유받은 링크에 답장시
-            if(matchedMemberId != null)
-            {
+            if (matchedMemberId != null) {
                 //상대방 답변 조회
                 List<AnswerDto.AnswerList> matchedMemberAnswerList = answerRepository.findAnswerList(matchedMemberId, null, categoryId);
 
@@ -66,10 +60,11 @@ public class MemberAnswerService
                 resultMap.put("matchedMemberName", memberRepository.findById(matchedMemberId).get().getName());
             }
 
+            //답변 리스트 resultMap에 저장
             resultMap.put("memberAnswerList", memberAnswerListMap);
 
             return resultMap;
-        }catch (Exception e) {
+        } catch (Exception e) {
             log.error("내 답변 리스트들을 불러오던 중 오류가 발생했습니다.", e);
             throw new CustomBadRequestException("내 답변 리스트들을 불러오던 중 오류가 발생했습니다.");
         }
@@ -77,9 +72,8 @@ public class MemberAnswerService
 
     //답안 저장
     @Transactional
-    public void saveAnswerList(MemberAnswerDto.Request memberAnswerDtoRequest)
-    {
-        try{
+    public void saveAnswerList(MemberAnswerDto.Request memberAnswerDtoRequest) {
+        try {
             //답안 파싱
             List<Long> answerList = memberAnswerDtoRequest.getAnswerList();
 
@@ -90,16 +84,19 @@ public class MemberAnswerService
             //카테고리 ID 파싱
             Long categoryId = memberAnswerDtoRequest.getCategoryId();
 
+            if (categoryRepository.findById(categoryId).isEmpty()) {
+                throw new CustomBadRequestException("카테고리 ID가 올바르지 않습니다!");
+            }
+
             //기존 답변목록 조회 및 삭제
             List<MemberAnswer> oldMemberAnswerList = memberAnswerRepository.findMemberAnswer(memberId, matchedMemberId, categoryId);
 
-            if(!oldMemberAnswerList.isEmpty() || oldMemberAnswerList != null)
-            {
+            if (!oldMemberAnswerList.isEmpty() || oldMemberAnswerList != null) {
                 //삭제대상 IdList
                 List<Long> deleteIdList = oldMemberAnswerList
-                                                            .stream()
-                                                            .map(o -> o.getMemberAnswerId())
-                                                            .collect(Collectors.toList());
+                        .stream()
+                        .map(o -> o.getMemberAnswerId())
+                        .collect(Collectors.toList());
 
                 //삭제
                 memberAnswerRepository.deleteAllById(deleteIdList);
@@ -107,54 +104,53 @@ public class MemberAnswerService
 
             //새로운 MemberAnswer 리스트 생성
             List<MemberAnswer> memberAnswer = answerList
-                                    .stream()
-                                    .map(a -> memberAnswerMapper.map(a, categoryId, memberId, matchedMemberId, new Date(), new Date()))
-                                    .collect(Collectors.toList());
+                    .stream()
+                    .map(a -> memberAnswerMapper.map(a, categoryId, memberId, matchedMemberId, new Date(), new Date()))
+                    .collect(Collectors.toList());
 
             //DB 저장
             memberAnswerRepository.saveAll(memberAnswer);
-        }catch (Exception e) {
+        } catch (CustomBadRequestException e) {
+            throw e;
+        } catch (Exception e) {
             log.error("답변을 저장하던중 오류가 발생했습니다.", e);
             throw new CustomBadRequestException("답변을 저장하던중 오류가 발생했습니다.");
         }
     }
 
-    public List<MatchedDto> matchedMemberList(Long categoryId)
-    {
-        try{
+    @Transactional(readOnly = true)
+    public List<MatchedDto> matchedMemberList(Long categoryId) {
+        try {
             Long myMemberId = Long.parseLong(SessionUtil.getUser().getUsername());
 
             //나와 매칭된 유저 목록 조회
             List<MatchedDto> matchedMemberList = memberAnswerRepository.findMatchedMemberList(categoryId, myMemberId);
 
-            if(!matchedMemberList.isEmpty() && matchedMemberList != null)
-            {
+            if (!matchedMemberList.isEmpty() && matchedMemberList != null) {
                 //내 답변 조회
                 List<AnswerList> myAnswerList = answerRepository.findAnswerList(myMemberId, null, categoryId);
 
                 //matchScore 계산
                 matchedMemberList
-                    .stream()
-                    .forEach(m -> m.setMatchScore(getMatchScore(myAnswerList, answerRepository.findAnswerList(m.getMemberId(), myMemberId, categoryId))));
+                        .stream()
+                        .forEach(m -> m.setMatchScore(getMatchScore(myAnswerList, answerRepository.findAnswerList(m.getMemberId(), myMemberId, categoryId))));
             }
 
             return matchedMemberList;
-        }catch (Exception e){
+        } catch (Exception e) {
             log.error("유저 목록을 불러오던 중 오류가 발생했습니다.", e);
             throw new CustomBadRequestException("유저 목록을 불러오던 중 오류가 발생했습니다.");
         }
     }
 
-    public double getMatchScore(List<AnswerList> memberAnswerList, List<AnswerList> matchedMemberAnswerList)
-    {
+    public double getMatchScore(List<AnswerList> memberAnswerList, List<AnswerList> matchedMemberAnswerList) {
         double sum = 0;
 
-        for(int i = 0; i < memberAnswerList.size(); i++)
-        {
+        for (int i = 0; i < memberAnswerList.size(); i++) {
             Long memberAnswerId = memberAnswerList.get(i).getAnswerId();
             Long matchedMemberAnswerId = matchedMemberAnswerList.get(i).getAnswerId();
 
-            if(memberAnswerId == matchedMemberAnswerId)
+            if (memberAnswerId == matchedMemberAnswerId)
                 sum++;
         }
